+++ src-tauri/Cargo.toml
@@ -24,7 +24,7 @@ tauri = { version = "2", features = [ "tray-icon", "macos-private-api"] }
 tauri-plugin-opener = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-bollard = "0.18"
+bollard = { version = "0.18", features = ["pipe"] }
 dirs = "5"
 image = "0.24"
 futures-util = "0.3"
